Fix the resultant path
The vertex of origin wasn't beign marked as visited but rather just the first value  in the visited list wich produced the wrong resultant path. Also this code fix another problem: the resultant path being produced by the original code that simply disappered with the first vertex (visited[0]), and also repeated a vertex wich is wrong by the definition of what is the christofides circuit(no repeting vertex with exception of the initial vertex that will appear in the first and last position).